// InvoiceSync-fe/src/app/core/models/page-response-company-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CompanyResponseDto } from './company-response-dto';
export interface PageResponseCompanyResponseDto {
  content: Array<CompanyResponseDto>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}